+++ src/main/java/net/oneandone/lavender/cli/Svn.java
@@ -20,7 +20,7 @@ import net.oneandone.lavender.config.Cluster;
 import net.oneandone.lavender.config.Docroot;
 import net.oneandone.lavender.config.Pool;
 import net.oneandone.lavender.config.SystemProperties;
-import net.oneandone.lavender.index.Distributor;
+import net.oneandone.lavender.modules.Distributor;
 import net.oneandone.lavender.index.Index;
 import net.oneandone.lavender.modules.Module;
 import net.oneandone.lavender.modules.SvnProperties;