+++ src/main/java/net/oneandone/lavender/cli/War.java
@@ -19,7 +19,7 @@ import net.oneandone.lavender.config.Cluster;
 import net.oneandone.lavender.config.Docroot;
 import net.oneandone.lavender.config.Pool;
 import net.oneandone.lavender.config.SystemProperties;
-import net.oneandone.lavender.index.Distributor;
+import net.oneandone.lavender.modules.Distributor;
 import net.oneandone.sushi.fs.file.FileNode;
 import net.oneandone.sushi.xml.XmlException;
 import org.xml.sax.SAXException;